Finish implementing Initial Condition
Create the initialcondition class and the way from Configuration to
Simulation. Tests has to made yet.

+++ Project/src/algorithm/DataFluid.java
@@ -3,7 +3,7 @@ package algorithm;
 public final class DataFluid {
 	
 	/** Fração de água no líquido */
-	public int BSW;
+	public float BSW;
 	/** Grau API do óleo do reservatório */
 	public double APi;
 	/** Massa específica do gás */
